requirements.txt
torch
autodistill
numpy>=1.20.0
opencv-python>=4.6.0
rf_groundingdino
rf_segment_anything
supervision
